refactor: move assets
