+++ src/assets/images/category/furniture.webp
