+++ src/assets/images/category/index.js
