+++ src/assets/images/sample/index.js
