+++ src/components/shared/thumbnailContainer/houseContainer/HouseContainer.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { StImgWrap, StThumbnailContainer } from "../styled";
 import Thumbnail from "../../thumbnail/Thumbnail";
 import NumberComma from "../../numberComma/NumberComma";
-import { home_img_01 } from "../../../../asset/images/sample";
+import { home_img_01 } from "../../../../assets/images/sample";
 
 function HouseContainer({ articleList, column }) {
   const navigate = useNavigate();
